Create Token class
The Token class represents a single immutable lexer token. Implemented
using a struct, you aren't meant to directly access the fields inside it.
Instead, use the functions defined below it.

There's also an enum associated with one of the Token members, called
TokenType. This defines what type of Token is represented.

A TokenType to string converter was also made, which is helpful for debug or error messages within the lexer.

The TokenType_punctuator function is like TokenType_toString but it does the opposite, giving you the correct TokenType based on the input.