Add name check to match()
The match() function wasn't entirely complete, since it didn't match a name against the lexer rules. It's perfectly reasonable to be able to do so since the lexer rules are so simple that you don't need any lookahead.

This also allowed me to refactor the Lexer_next() algorithm to remove an extra brace nest.