Replace names with integers
The Lexer code was simplified and refactored by only being able to parse
numbers and symbols rather than variable names, which would've required a
substring.

In the Token struct, the begin and end pointers were replaced by a long int to
represent the value of the token, like for an integer token.

A test was written in main.c to verify the functionality of the Lexer.