Remove destructor from Expr vtable
The Expr class is what's called an abstract class. For that to happen, it
needs to have a pure virtual destructor. Setting its destructor in the vtable
to NULL accomplishes that by not allowing an instance of an abstract class to
be created/destroyed in a valid way, forcing a segfault.