Add prefix parselets
The PrefixParselet abstract base class is a type of parselet built for parsing
things like unary operators, parenthesis groups, and integers. It uses a
virtual call to its parse() function, taking the Parser used and the Token
that caused the Parser to call that particular parselet.

The parse() functions currently have a placeholder returning NULL. They're just
here to provide a useful interface to the Parser before the parselets are
actually implemented.

There are 3 PrefixParselet subclasses:

The GroupParselet parses an expression grouped within parentheses, e.g.
"(a+b)". A placeholder was put inside its overriden v_parse() function for now.

The IntParselet is for parsing integer expressions, i.e. a single integer
token. It really doesn't do much else here but it's there for parsing purposes.

The UnaryParselet is for parsing unary operators like "-a". Since operators
have precedence, it also has a member variable called precedence.