Add infix parselets
InfixParselets are for expressions containing an infix operator, i.e. an
operator that comes right after the expression it affects. An example of an
infix operator would be the "+" in "a+b". Parselets like these also include
some types of unary operators that come after the expression, like the "!" in
"a!".

There is only one subclass for now: BinaryParselet.

The BinaryParselet is for parsing typical binary operators like "a+b". Since
different operators have different precedences and associativities, they were
included in the constructor for use in the parse() function. That way, I won't
have to write tons of different parselets for every single operator.