Add lookahead support
There are some cases where the parser needs to look at the next token that the
lexer has, but not consuming it just yet. To do that, a member function (and
variable) was added to the Parser class. Its pre-existing consume() function
was modified to use the new method of getting Tokens from the Lexer.