Add virtual functions to Parser
Before the parser is implemented, it needs a way to use the parselets. To do
that, I could use a switch statement with a TokenType to get the right prefix
or infix parselet, but what if I want to extend this Parser for another
grammar? Instead of modifying the Parser class itself, a more object-oriented
alternative would be to use a virtual call.