Fix possible access violation
Everytime Lexer_next is called, it consumes the next character of the string.
When the lexer already hit the end of the string, what happens then? It still
consumes the next char after that! An if statement was put around the entire
input loop so that it will always return an eof token when at the end of the
input.