Use a more efficient exponentiation algorithm
The previous one used the <math.h> pow(), which took doubles as arguments. This
can be a problem, because ints don't translate well into doubles. An example is
the expression "10^2" being 99 instead of 100. The algorithm I found online,
which is an iterative approach that takes O(1) space and O(log(n)) time.